services/technical-master-ai/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
google-auth==2.35.0
google-cloud-bigquery==3.27.0
google-api-core==2.22.0
